requirements.txt
Django==1.11.24
django-tastypie==0.14.2
django-cors-headers==3.1.1
gunicorn==19.9.0
requests==2.20.0
psycopg2-binary==2.8.4
pycryptodome==3.9.0